Reorganize code in separate modules
