+++ src/app/game/game-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { GameComponent } from 'src/app/game/game.component';
+
+const routes: Routes = [{ path: '', component: GameComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class GameRoutingModule {}
