+++ src/app/game/question-points/question-points.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { QuestionPointsComponent } from 'src/app/game/question-points/question-points.component';
+
+describe('QuestionPointsComponent', () => {
+  let component: QuestionPointsComponent;
+  let fixture: ComponentFixture<QuestionPointsComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [QuestionPointsComponent],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(QuestionPointsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
