--- src/app/level-selection/level-selection.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { LevelSelectionComponent } from './level-selection.component';
-
-describe('LevelSelectionComponent', () => {
-  let component: LevelSelectionComponent;
-  let fixture: ComponentFixture<LevelSelectionComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [LevelSelectionComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(LevelSelectionComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
