+++ src/app/pregame/player-selection/player-selection.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { PlayerSelectionComponent } from 'src/app/pregame/player-selection/player-selection.component';
+
+describe('PlayerSelectionComponent', () => {
+  let component: PlayerSelectionComponent;
+  let fixture: ComponentFixture<PlayerSelectionComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [PlayerSelectionComponent],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(PlayerSelectionComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
