+++ src/app/pregame/pregame-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { PregameComponent } from 'src/app/pregame/pregame.component';
+
+const routes: Routes = [{ path: '', component: PregameComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class PregameRoutingModule {}
